requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pillow==10.3.0
numpy==1.26.4
